feat: adiciona addObjective useCase e CultivationTests

+++ backend/Src/Application/Src/UseCases/User/DeleteUser/DeleteUser.cs
@@ -1,5 +1,4 @@
 using Aquadata.Application.Interfaces;
-using Aquadata.Application.UseCases.Pond.Common;
 using Aquadata.Core.Interfaces.Repository;
 using Aquadata.Core.Util;
 using Aquadata.Core.Util.Result;
@@ -10,12 +9,15 @@ namespace Aquadata.Application.UseCases.User.DeleteUser;
 public class DeleteUser : IUseCaseHandler<DeleteUserInput,Unit>
 {
   private readonly IUserRepository _userRepository;
+  private readonly IAuthenticatedUserService _authenticatedUserService;
   private readonly IUnitOfWork _unitOfWork;
 
-  public DeleteUser(IUserRepository userRepository, IUnitOfWork unitOfWork)
+  public DeleteUser(IUserRepository userRepository, 
+  IUnitOfWork unitOfWork, IAuthenticatedUserService authenticatedUserService)
   {
     _userRepository = userRepository;
     _unitOfWork = unitOfWork;
+    _authenticatedUserService = authenticatedUserService;
   }
 
   public async Task<Result<Unit>> Handle(DeleteUserInput request, 
@@ -30,6 +32,15 @@ public class DeleteUser : IUseCaseHandler<DeleteUserInput,Unit>
           "User not found'"
         )
       );
+    
+    var userId = _authenticatedUserService.GetUserId();
+    if (userId != user.Id.ToString())
+      return Result<Unit>.Fail(
+        Error.Unauthorized(
+          "UseCases.Pond.Deactivate",
+          "Unauthorized"
+        )
+    );
 
     await _userRepository.Delete(user, cancellationToken);
     await _unitOfWork.Commit(cancellationToken);
